+ Specify configuration options in extra config file
By default, expsrc.sh searches in the project to export root directory for the file "expsrc.cfg" if found, it reads the file and applies the configuration as specified in this file.

This can be used typically to define patterns for parsing files.

The syntax in the file is one command line option per line. Comments are allowed, a leading '#' marks a comment for a line.

If an alternative configuration name is used, the option --config can be used to specify the location of the config file.